Restore coast mode in a dedicated thread.
Coast mode (where the drive and steer motors will be in coast mode when the robot is disabled) is useful for moving the robot around the field. This was problematic in 2024 since changing the configuration of a CTRE device may take a significant amount of time (~200 ms). We now change the neutral mode configuration in a separate thread.